+++ SimulationTest/cpuTests/CollisionTest.cpp
@@ -8,7 +8,7 @@
 #include <algorithm>
 
 template <typename T>
-bool sortVector(T const &lhs, T const &rhs, std::vector<double*> compareLhs, std::vector<double*> compareRhs) {
+bool sortVector(T const &lhs, T const &rhs, std::vector<float*> compareLhs, std::vector<float*> compareRhs) {
 	if (*(compareLhs.back()) < *(compareRhs.back())) {
       	return true; 
 	} else if (*(compareRhs.back()) < *(compareLhs.back())) {
@@ -24,7 +24,7 @@ bool sortVector(T const &lhs, T const &rhs, std::vector<double*> compareLhs, std
 }
 
 bool sortParticles(Particle* const &lhs, Particle* const &rhs) {
-	auto compareLhs = std::vector<double*> {
+	auto compareLhs = std::vector<float*> {
 		&(lhs->velocity.z),
 		&(lhs->velocity.y),
 		&(lhs->velocity.x),
@@ -34,7 +34,7 @@ bool sortParticles(Particle* const &lhs, Particle* const &rhs) {
 		&(lhs->radius),
 		&(lhs->mass)
 	};
-	auto compareRhs = std::vector<double*> {
+	auto compareRhs = std::vector<float*> {
 		&(rhs->velocity.z),
 		&(rhs->velocity.y),
 		&(rhs->velocity.x),
