+++ SimulationTest/cpuTests/DistributionSimpleTest.cpp
@@ -3,7 +3,7 @@
 
 TEST(DistributionSimpleTest, DistributionSimpleTest) {
 	DistributionSimple distribution(2, 1);
-	double value = distribution.getValue();
+	float value = distribution.getValue();
 	EXPECT_TRUE(value >= 1 && value <= 3);
 }
